requirements.txt
python-telegram-bot==21.0.1
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
aiohttp==3.11.7